Make it easier to add tests
Instead of having to deal with a massive mess of rules every time you
want to add a new test, you can now just add a new .spec.ts file (as
long as it is in generic_ui/scripts or generic_core).

This change looks a lot scarier than it is, so a quick description of
what it really done:

This moves the gruntConfig object to a variable named gruntConfig

This adds a new function to the Gruntfile to look for all the spec files
in a directory and make rules to browserify those files and run them
(return the tasks to run to do that).

This uses that function for generic_ui/scripts and generic_core

That's all!  Sorry about how ugly the diff is going to look...