Make it easier to register commands
This makes it easier to register commands by adding an array we loop
through.

In order to support that, we needed a single function we could call in
order to register a function that may or may not return a promise, this
is done as a new onCommand function.

In order to have a single function, we now have the onCommand do error
checking to determine whether a promise was passed/expected and then
have it send the promise.  Anyways, that's going to take error handling.
Yay!

Because that is confusing, this also adds tests.

Review for this should include answering the question: Does this
actually improve anything?