Add a new field, hasSeenMetrics, use it to keep track of that
Right now we were using hasSeenWelcome to determine whether the user had
seen the welcome screen and also refer to it when deciding whether to
show the screen for opting-in to metrics.  This caused you to see it
multiple times when hitting back during the first run.  This fixes that.