Move some of the files in the dist copy arrays to common locations
We were previously doing a LOT of duplicate copying for the builds on
different platforms.  With this change, a lot of those files will only
be specified once

This also fixes an error with a missing file in cordova as well as an
extra file that should have been ignored.