Remove some duplicated file paths from Gruntfile
In the Gruntfile, for almost every platform, we have some rules to copy
specific node modules due to freedom module requirements.  This changes
those rules around a bit so that we now just build a list of all the
node modules we want to copy in that way and then use it.