Allow the use of a proxy bypass on Chrome
This adds the ability for the user to specify a list of addresses that
will not be sent through uProxy.  By default, this is set to local
addresses.

For now, this is only working on Chrome since matching the URI on
Firefox will need to be done manually.

Note that updating the addresses while a proxying session is in progress
will not currently actually do anything.  The addresses are only checked
when a connection is started.