Add a component to live in the real background
This adds a BackgroundUI class that is intended to live in the real
background context instead of the UI object we have now which hovers
between the panel context and the background depending on how we need to
use it.  In the coming week, I will be sending additional pull requests
which will move components out of ui.ts to either background_ui.ts or a
Polymer component depending on where that code should actually live.

This is intended to move us in the direction of being able to transition
from Polymer 0.5 to 1.0 as we will no longer be able to rely on
data-sharing between objections inside of and outside of Polymer to
propogate state changes.

This also includes proof-of-concept components that handle message
passing between the panel and BackgroundUI (and a transitionary way
of allowing the current UserInterface to listen here).