Move a lot of the dialog code out of the background
So, a bunch of changes.

First of all, in the panel, we now no longer make calls directly on the
UI object to open panels.  This code is handled completely in the panel
with message passing, so this is a nice gain in removing those
round-trips.

At the same time, this also pulls out a bunch of the code from the UI
that ended up opening the dialogs (which were only accessed in the
panel) directly into the panel.

For the dialogs, this also removes the fire-signal property which
over-complicated things.  Instead, elements that want to fire a signal
upon completion just register a promise handler there.

Finally, this moves the code to create the dialog box descriptions to a
common util file that is used in both the panel and in the background.