Ditch "Loading uProxy friends" screen.
Was being displayed unconditionally for 5 seconds, even when it shouldn't have
(e.g. it's the first time the user is running uProxy and they don't yet have
any friends to load, or the user is not currently signed into any networks so
there's actually no loading activity going on, etc).

This is extra problematic because the "Loading uProxy friends" screen covers
up the + button. So every time a user opened uProxy with the goal of adding a
friend (remembering they might not yet even be signed into any networks),
they'd first have to sit through a 5-second "loading" screen (that possibly
advertised background activity that wasn't even really happening and) that was
only getting in the way of accomplishing the goal.

It looks like this used to make more sense (e.g. in 99cf70dd), but the contacts
logic changed multiple times since then, and the soundness of the UI logic
didn't keep pace. Seems strictly better to remove the current, problematic
"loading friends" UI for now, and just have contacts appear once they're
loaded. A more sound "loading friends" UI can be added back in a follow-up
change if deemed necessary by proper design review.

Closes #2417
Closes #2418