Switch to getting port control status in the UI
This switches the UI over to getting the port control status on the
advanced-settings page instead of the previous behaviour where we would
have a globally-cached value that the UI panel would merely access.

We load this value the first time the page is opened, it will stay
cached throughout the lifetime of the advanced-settings page (i.e.
lifetime of the panel).