Remove splashState from global settings
The only reason we would want to keep splashState in globalSettings is
to ensure that if the user closes uproxy while on the metrics screen,
they are taken back to the metrics screen instead of the screen before
it.  That takes a lot of overhead to do, so this change removes that
functionality.

This does theoretically break backwards-compatibility between the app
and the extension, however, since splashState would be in storage for
anyone who had just upgraded the app, that's even more unlikely than
usual to come up.

This is in support of the #2556 and indirectly #1900 (Polymer 1.0)