Remove global data dependencies from roster.*
This removes the dependencies on data from ui and model in roster.html
and roster.ts.

In roster.html, we are able to remove the dependency on
model.globalSettings.mode by instead looking at the mode attribute given
to the element (which mode it is for).  We can do this because it's fine
for the get roster list to be displaying "OFFERS" even when it's hidden.

In roster.ts, we can remove the dependencies on looking at the contacts
by instead just looking at the contacts specifically passed into this
roster object.

This also removes the (now unused) loadingContacts variable.

This is in support of #2556 and indirectly #1900 (Polymer 1.0)