Stop using state sharing for gettingStatus/sharingStatus
When updating gettingStatus and sharingStatus in the UI, we now send a
signal noting that they have changed instead of our previous behaviour
of just updating shared variables.

It seems very possible that we could reduce the complexity further by
looking at more underlying data (e.g. instanceGettingAccessFrom,
offeringInstances) but that seems unnecessary for this project.

This is in support of #2556 and indirectly #1900 (Polymer 1.0)