Add sane behaviour to the Android back button
To start with, we add, to most of the views that include a
panel/overlay/dialog, an `opened` property that simply states whether
that panel is active and a `close` method that closes the panel.

For the `invite-user` view, we use the method back instead of close.  In
this case, there are multiple different panels that it might have open
instead of being the main panel itself and we want to make sure we close
the correct (active) one only.

In order to handle the overall functionality of the back button, we now
listen in root.ts for the backbutton core signal.  Our handler there
then cycles through the possible things that could be the top-level
element and closes whichever one that is.  If there is nothing else
open, we close the entire app with a new method that's been added to the
browser api.

Also, as a result of this change, a lot of elements that previously
handled sending open events to their direct children now do so via just
calling the open method on the child element.  It's a lot cleaner and
will hopefully remove some (rarely-suffered) overhead.

Fixes #2440