Send an update to the UI if global settings are changed
If the global settings change from an action that is not taken within
the UI, we should send an update to the UI with a new copy of
globalSettings.  In this case, we're sending the entire object since we
are treating the core version of globalSettings as the canonical one.